12/6 finish handler modules and create util
