+++ server.js
@@ -8,6 +8,8 @@ import createRoomHandler from "./src/handlers/createRoom.js";
 import joinRoomHandler from "./src/handlers/joinRoom.js";
 import voteSocketHandler from "./src/handlers/voteSocket.js";
 import showVotesHandler from "./src/handlers/showVotes.js";
+import leaveRoomHandler from "./src/handlers/leaveRoom.js";
+import disconnectSocketHandler from "./src/handlers/disconnectSocket.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -70,117 +72,9 @@ io.on("connection", async (socket) => {
   joinRoomHandler(io, socket, sharedData);
   voteSocketHandler(io, socket, sharedData);
   showVotesHandler(io, socket, sharedData);
-
-  socket.on("disconnectButtonAction", (arg, callback) => {
-    const roomName = Array.from(socket.rooms)[1];
-    const socketId = socket.id;
-    const socketName = socketNamesAndSocketIDs.get(socketId);
-    console.log("roomsData[roomName]");
-    console.log(roomsData[roomName]);
-    console.log("socketName");
-    console.log(socketName);
-    const toBeDeletedIndex = roomsData[roomName].findIndex(
-      (s) => s === socketName
-    );
-    roomsData[roomName].splice(toBeDeletedIndex, 1);
-    votesData[roomName].splice(toBeDeletedIndex, 1);
-    socketNamesAndSocketIDs.delete(socket.id);
-    socket.leave(roomName);
-
-    if (roomsData[roomName].length < 1) {
-      delete roomsData[roomName];
-      delete votesData[roomName];
-    }
-
-    io.to(roomName).emit("newSocketInRoom", roomsData[roomName]);
-    io.to(roomName).emit("showVotesFromServer", votesData[roomName]);
-    callback({
-      toBeDeletedIndex,
-      roomsData,
-      votesData,
-    });
-  });
-
-  socket.on("disconnect", () => {
-    console.log("socket.id");
-    console.log(socket.id);
-    const socketId = socket.id;
-    let socketName;
-    if (socketNamesAndSocketIDs.get(socketId)) {
-      socketName = socketNamesAndSocketIDs.get(socketId);
-
-      //GPT
-      // Example: Find the string "dog" in the object of arrays
-      let searchString = socketName;
-      // Example: Find the string "dog" in the object of arrays
-      let roomName = findStringInObject(roomsData, searchString);
-
-      if (roomName !== null) {
-        console.log(
-          `Found "${searchString}" in the array with property name "${roomName}".`
-        );
-      } else {
-        console.log(`"${searchString}" not found in any array.`);
-      }
-
-      console.log("socketName");
-      console.log(socketName);
-
-      console.log("roomName");
-      console.log(roomName);
-
-      const toBeDeletedIndex = roomsData[roomName].findIndex(
-        (s) => s === socketName
-      );
-      roomsData[roomName].splice(toBeDeletedIndex, 1);
-      votesData[roomName].splice(toBeDeletedIndex, 1);
-      socketNamesAndSocketIDs.delete(socket.id);
-      socket.leave(roomName);
-
-      if (roomsData[roomName].length < 1) {
-        delete roomsData[roomName];
-        delete votesData[roomName];
-      }
-
-      io.to(roomName).emit("newSocketInRoom", roomsData[roomName]);
-      io.to(roomName).emit("showVotesFromServer", votesData[roomName]);
-      console.log("socket disconnect data clear successful");
-    } else {
-      console.log("no name or rooms associated with the socket");
-    }
-  });
+  leaveRoomHandler(io, socket, sharedData);
+  disconnectSocketHandler(io, socket, sharedData);
 });
-
-// Function to find a string in the object of arrays and return the property name
-function findStringInObject(object, searchString) {
-  let foundKey = null;
-
-  // Iterate through the keys (property names) of the object
-  for (let key in object) {
-    if (
-      Object.prototype.hasOwnProperty.call(object, key) &&
-      Array.isArray(object[key])
-    ) {
-      // Iterate through the elements of each array
-      for (let i = 0; i < object[key].length; i++) {
-        // Check if the array element is a string and contains the search string
-        if (
-          typeof object[key][i] === "string" &&
-          object[key][i].includes(searchString)
-        ) {
-          foundKey = key;
-          break; // Stop iterating if a match is found
-        }
-      }
-    }
-
-    if (foundKey) {
-      break; // Stop iterating if a match is found in any array
-    }
-  }
-
-  return foundKey;
-}
 
 server.listen(10000, () => {
   console.log("server running at http://localhost:10000");
