+++ src/handlers/disconnectSocket.js
@@ -1,22 +1,54 @@
-export default function createRoomHandler(io, socket, data) {
-  socket.on("socketCreateRoom", (arg, callback) => {
-    if (arg) {
-      const socketName = arg;
-      data.socketNamesAndSocketIDs.set(socket.id, socketName);
-      const currentRoomName = "room" + socket.id;
-      socket.join(currentRoomName);
-      data.roomsData[currentRoomName] = [socketName];
-      data.votesData[currentRoomName] = ["0"];
+import { findStringInObject } from "../util/findStringInObject.js";
 
-      io.to(currentRoomName).emit(
-        "newSocketInRoom",
-        data.roomsData[currentRoomName]
+export default function disconnectSocketHandler(io, socket, data) {
+  const { socketNamesAndSocketIDs, roomsData, votesData } = data;
+
+  socket.on("disconnect", () => {
+    console.log("socket.id");
+    console.log(socket.id);
+    const socketId = socket.id;
+    let socketName;
+    if (socketNamesAndSocketIDs.get(socketId)) {
+      socketName = socketNamesAndSocketIDs.get(socketId);
+
+      //GPT
+      // Example: Find the string "dog" in the object of arrays
+      let searchString = socketName;
+      // Example: Find the string "dog" in the object of arrays
+      let roomName = findStringInObject(roomsData, searchString);
+
+      if (roomName !== null) {
+        console.log(
+          `Found "${searchString}" in the array with property name "${roomName}".`
+        );
+      } else {
+        console.log(`"${searchString}" not found in any array.`);
+      }
+
+      console.log("socketName");
+      console.log(socketName);
+
+      console.log("roomName");
+      console.log(roomName);
+
+      const toBeDeletedIndex = roomsData[roomName].findIndex(
+        (s) => s === socketName
       );
+      roomsData[roomName].splice(toBeDeletedIndex, 1);
+      votesData[roomName].splice(toBeDeletedIndex, 1);
+      socketNamesAndSocketIDs.delete(socket.id);
+      socket.leave(roomName);
 
-      console.log(data.roomsData);
-      callback(currentRoomName);
+      if (roomsData[roomName].length < 1) {
+        delete roomsData[roomName];
+        delete votesData[roomName];
+      }
+
+      io.to(roomName).emit("newSocketInRoom", roomsData[roomName]);
+      io.to(roomName).emit("showVotesFromServer", votesData[roomName]);
+      console.log("socket disconnect data clear successful");
     } else {
-      callback("error");
+      console.log("no name or rooms associated with the socket");
     }
   });
 }
